+++ packages/create-react-router/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `create-react-router`
 
+## 7.8.1
+
+_No changes_
+
 ## 7.8.0
 
 _No changes_